pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vvnext"
version = "0.1.0"
description = "sing-box multi-node proxy fleet manager"
requires-python = ">=3.10"
license = "Apache-2.0"
dependencies = [
    "typer>=0.12",
    "questionary>=2.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "paramiko>=3.0",
    "httpx>=0.27",
    "cryptography>=42.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov", "ruff", "mypy"]

[project.scripts]
vvnext = "vvnext.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/vvnext"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.pytest.ini_options]
testpaths = ["tests"]
